Mcc and mnc from xml in RRO may be integers
Mcc and mnc in an xml from overlays built in current android build
system are integers to the contrary of the apns-full-conf.xml
where they are strings. The telephony provider uses the same parsing
mechanism in both cases which makes the full apn list get
correct values whereas the provider misjudge what it should
insert in the database when there is an apns.xml in an overlay.

For the newly built android RROs mnc="01" becomes mnc="1" and no
match is found when the sim card is inserted.

To not risk the large APN list we only change the behavior for the
overlay handling.

Solution is to format the mcc value to have 3 digits and use the
carrier id table to get correct mnc. Sim cards missing
in the carrier id table need to be added there.

Test: Manual
Test: adb shell content query --uri content://telephony/carriers
Bug: 175620622
Change-Id: I244ad0f76f360a0635343187e455b54c90ec45b4